+++ c++_implementation/source/BayesClassifier.h
@@ -11,7 +11,7 @@ public:
 	
 	label_data_t predict(const image_data_t & test_data);
 
-	image_data_t generate(const short N, const short label);
+	image_data_t generate(const size_t N, const short label);
 	
 private:
 	//private variables
